requirements.txt
pycryptodome==3.8.2